APPLINK-4694: Fixed some flexelint warnings

+++ src/components/application_manager/src/commands/hmi/on_button_event_notification.cc
@@ -52,7 +52,7 @@ void OnButtonEventNotification::Run() {
   LOG4CXX_INFO(logger_, "OnButtonEventNotification::Run");
 
   (*message_)[strings::params][strings::function_id] =
-      mobile_apis::FunctionID::OnButtonEventID;
+      static_cast<int>(mobile_apis::FunctionID::OnButtonEventID);
 
   SendNotificationToMobile(message_);
 }
